requirements.txt
pyyaml
flask
flask_restful
GitPython
gunicorn
pylint